TESTS: move dataframe_backend fixture to conftest
